+++ playground/react-native-playground/CHANGELOG.md
@@ -7,6 +7,32 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.1]
+
+### Changed
+
+- Bump workspace dependencies:
+  - @metamask/connect-evm@0.9.0
+  - @metamask/connect-multichain@0.11.0
+
+## [0.3.0]
+
+### Added
+
+- Add `integrationType: wagmi` to wagmi metamask connector([#215](https://github.com/MetaMask/connect-monorepo/pull/215/))
+
+### Changed
+
+- Update `getInfuraRpcUrls` calls to use new options object parameter ([#211](https://github.com/MetaMask/connect-monorepo/pull/211))
+
+## [0.2.1]
+
+### Changed
+
+- Bump workspace dependencies:
+  - @metamask/connect-evm@0.7.0
+  - @metamask/connect-multichain@0.9.0
+
 ## [0.2.0]
 
 ### Changed
@@ -62,7 +88,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[Unreleased]: https://github.com/MetaMask/connect-monorepo/compare/@metamask/react-native-playground@0.2.0...HEAD
+[Unreleased]: https://github.com/MetaMask/connect-monorepo/compare/@metamask/react-native-playground@0.3.1...HEAD
+[0.3.1]: https://github.com/MetaMask/connect-monorepo/compare/@metamask/react-native-playground@0.3.0...@metamask/react-native-playground@0.3.1
+[0.3.0]: https://github.com/MetaMask/connect-monorepo/compare/@metamask/react-native-playground@0.2.1...@metamask/react-native-playground@0.3.0
+[0.2.1]: https://github.com/MetaMask/connect-monorepo/compare/@metamask/react-native-playground@0.2.0...@metamask/react-native-playground@0.2.1
 [0.2.0]: https://github.com/MetaMask/connect-monorepo/compare/@metamask/react-native-playground@0.1.3...@metamask/react-native-playground@0.2.0
 [0.1.3]: https://github.com/MetaMask/connect-monorepo/compare/@metamask/react-native-playground@0.1.2...@metamask/react-native-playground@0.1.3
 [0.1.2]: https://github.com/MetaMask/connect-monorepo/compare/@metamask/react-native-playground@0.1.1...@metamask/react-native-playground@0.1.2